WIP first pass reorg, untested
